v26.0.29496.0

+++ app/Base Application/codeunit/Whse-Activity-Register.C7307.al
@@ -73,6 +73,7 @@ codeunit 7307 "Whse.-Activity-Register"
         WhseInternalPickLine: Record "Whse. Internal Pick Line";
         WhseInternalPutAwayLine: Record "Whse. Internal Put-away Line";
         ProdCompLine: Record "Prod. Order Component";
+        ProdOrderLine: Record "Prod. Order Line";
         AssemblyLine: Record "Assembly Line";
         JobPlanningLine: Record "Job Planning Line";
         ProdOrder: Record "Production Order";
@@ -500,8 +501,13 @@ codeunit 7307 "Whse.-Activity-Register"
                 if (WhseActivLineGrouped."Action Type" <> WhseActivLineGrouped."Action Type"::Take) and (WhseActivLineGrouped."Breakbulk No." = 0) then
                     UpdateWhseIntPickLine(WhseActivLineGrouped);
             WhseActivLineGrouped."Whse. Document Type"::Production:
-                if (WhseActivLineGrouped."Action Type" <> WhseActivLineGrouped."Action Type"::Take) and (WhseActivLineGrouped."Breakbulk No." = 0) then
-                    UpdateProdCompLine(WhseActivLineGrouped);
+                if WhseActivLineGrouped."Source Document" = WhseActivLineGrouped."Source Document"::"Prod. Consumption" then begin
+                    if (WhseActivLineGrouped."Action Type" <> WhseActivLineGrouped."Action Type"::Take) and (WhseActivLineGrouped."Breakbulk No." = 0) then
+                        UpdateProdCompLine(WhseActivLineGrouped);
+                end else
+                    if WhseActivLineGrouped."Source Document" = WhseActivLineGrouped."Source Document"::"Prod. Output" then
+                        if (WhseActivLineGrouped."Action Type" <> WhseActivLineGrouped."Action Type"::Place) and (WhseActivLineGrouped."Breakbulk No." = 0) then
+                            UpdateProdOrderLine(WhseActivLineGrouped);
             WhseActivLineGrouped."Whse. Document Type"::Assembly:
                 if (WhseActivLineGrouped."Action Type" <> WhseActivLineGrouped."Action Type"::Take) and (WhseActivLineGrouped."Breakbulk No." = 0) then
                     UpdateAssemblyLine(WhseActivLineGrouped);
@@ -528,8 +534,8 @@ codeunit 7307 "Whse.-Activity-Register"
     /// <param name="WhseActivLine">Provided warehouse activity line.</param>
     procedure UpdateWhseDocHeader(WhseActivLine: Record "Warehouse Activity Line")
     var
-        WhsePutAwayRqst: Record "Whse. Put-away Request";
-        WhsePickRqst: Record "Whse. Pick Request";
+        WhsePutAwayRequest: Record "Whse. Put-away Request";
+        WhsePickRequest: Record "Whse. Pick Request";
         IsHandled: Boolean;
     begin
         IsHandled := false;
@@ -541,8 +547,7 @@ codeunit 7307 "Whse.-Activity-Register"
             WhseActivLine."Whse. Document Type"::Shipment:
                 if WhseActivLine."Action Type" <> WhseActivLine."Action Type"::Take then begin
                     WhseShptHeader.Get(WhseActivLine."Whse. Document No.");
-                    WhseShptHeader.Validate(
-                      "Document Status", WhseShptHeader.GetDocumentStatus(0));
+                    WhseShptHeader.Validate("Document Status", WhseShptHeader.GetShipmentStatus(0));
                     WhseShptHeader.Modify();
                 end;
             WhseActivLine."Whse. Document Type"::Receipt:
@@ -557,9 +562,9 @@ codeunit 7307 "Whse.-Activity-Register"
                     if PostedWhseRcptHeader."Document Status" =
                        PostedWhseRcptHeader."Document Status"::"Completely Put Away"
                     then begin
-                        WhsePutAwayRqst.SetRange("Document Type", WhsePutAwayRqst."Document Type"::Receipt);
-                        WhsePutAwayRqst.SetRange("Document No.", PostedWhseRcptHeader."No.");
-                        WhsePutAwayRqst.DeleteAll();
+                        WhsePutAwayRequest.SetRange("Document Type", WhsePutAwayRequest."Document Type"::Receipt);
+                        WhsePutAwayRequest.SetRange("Document No.", PostedWhseRcptHeader."No.");
+                        WhsePutAwayRequest.DeleteAll();
                         ItemTrackingMgt.DeleteWhseItemTrkgLines(
                           Database::"Posted Whse. Receipt Line", 0, PostedWhseRcptHeader."No.", '', 0, 0, '', false);
                     end;
@@ -605,24 +610,43 @@ codeunit 7307 "Whse.-Activity-Register"
                     end;
                 end;
             WhseActivLine."Whse. Document Type"::Production:
-                if WhseActivLine."Action Type" <> WhseActivLine."Action Type"::Take then begin
-                    ProdOrder.Get(WhseActivLine."Source Subtype", WhseActivLine."Source No.");
-                    ProdOrder.CalcFields("Completely Picked");
-                    if ProdOrder."Completely Picked" then begin
-                        WhsePickRqst.SetRange("Document Type", WhsePickRqst."Document Type"::Production);
-                        WhsePickRqst.SetRange("Document No.", ProdOrder."No.");
-                        WhsePickRqst.ModifyAll("Completely Picked", true);
-                        ItemTrackingMgt.DeleteWhseItemTrkgLines(
-                          Database::"Prod. Order Component", WhseActivLine."Source Subtype", WhseActivLine."Source No.", '', 0, 0, '', false);
-                    end;
+                begin
+                    if WhseActivLine."Source Document" = WhseActivLine."Source Document"::"Prod. Consumption" then
+                        if WhseActivLine."Action Type" <> WhseActivLine."Action Type"::Take then begin
+                            ProdOrder.Get(WhseActivLine."Source Subtype", WhseActivLine."Source No.");
+                            ProdOrder.CalcFields("Completely Picked");
+                            if ProdOrder."Completely Picked" then begin
+                                WhsePickRequest.SetRange("Document Type", WhsePickRequest."Document Type"::Production);
+                                WhsePickRequest.SetRange("Document No.", ProdOrder."No.");
+                                WhsePickRequest.ModifyAll("Completely Picked", true);
+                                ItemTrackingMgt.DeleteWhseItemTrkgLines(
+                                  Database::"Prod. Order Component", WhseActivLine."Source Subtype", WhseActivLine."Source No.", '', 0, 0, '', false);
+                            end;
+                        end;
+                    if WhseActivLine."Source Document" = WhseActivLine."Source Document"::"Prod. Output" then
+                        if WhseActivLine."Action Type" <> WhseActivLine."Action Type"::Place then begin
+                            ProdOrder.Get(WhseActivLine."Source Subtype", WhseActivLine."Source No.");
+                            ProdOrder."Document Put-away Status" := ProdOrder.GetHeaderStatus(0);
+                            ProdOrder.Modify();
+
+                            if ProdOrder."Document Put-away Status" =
+                               ProdOrder."Document Put-away Status"::"Completely Put Away"
+                            then begin
+                                WhsePutAwayRequest.SetRange("Document Type", WhsePutAwayRequest."Document Type"::Production);
+                                WhsePutAwayRequest.SetRange("Document No.", ProdOrder."No.");
+                                WhsePutAwayRequest.DeleteAll();
+                                ItemTrackingMgt.DeleteWhseItemTrkgLines(
+                                  Database::"Prod. Order Line", 0, ProdOrder."No.", '', 0, 0, '', false);
+                            end;
+                        end;
                 end;
             WhseActivLine."Whse. Document Type"::Assembly:
                 if WhseActivLine."Action Type" <> WhseActivLine."Action Type"::Take then begin
                     AssemblyHeader.Get(WhseActivLine."Source Subtype", WhseActivLine."Source No.");
                     if AssemblyHeader.CompletelyPicked() then begin
-                        WhsePickRqst.SetRange("Document Type", WhsePickRqst."Document Type"::Assembly);
-                        WhsePickRqst.SetRange("Document No.", AssemblyHeader."No.");
-                        WhsePickRqst.ModifyAll("Completely Picked", true);
+                        WhsePickRequest.SetRange("Document Type", WhsePickRequest."Document Type"::Assembly);
+                        WhsePickRequest.SetRange("Document No.", AssemblyHeader."No.");
+                        WhsePickRequest.ModifyAll("Completely Picked", true);
                         ItemTrackingMgt.DeleteWhseItemTrkgLines(
                           Database::"Assembly Line", WhseActivLine."Source Subtype", WhseActivLine."Source No.", '', 0, 0, '', false);
                     end;
@@ -632,9 +656,9 @@ codeunit 7307 "Whse.-Activity-Register"
                     Job.Get(WhseActivLine."Whse. Document No.");
                     Job.CalcFields("Completely Picked");
                     if Job."Completely Picked" then begin
-                        WhsePickRqst.SetRange("Document Type", WhsePickRqst."Document Type"::Job);
-                        WhsePickRqst.SetRange("Document No.", Job."No.");
-                        WhsePickRqst.ModifyAll("Completely Picked", true);
+                        WhsePickRequest.SetRange("Document Type", WhsePickRequest."Document Type"::Job);
+                        WhsePickRequest.SetRange("Document No.", Job."No.");
+                        WhsePickRequest.ModifyAll("Completely Picked", true);
                         ItemTrackingMgt.DeleteWhseItemTrkgLines(
                           Database::"Job Planning Line", WhseActivLine."Source Subtype", WhseActivLine."Source No.", '', 0, 0, '', false);
                     end;
@@ -677,7 +701,7 @@ codeunit 7307 "Whse.-Activity-Register"
 
         WhseShptLine.Validate("Qty. to Ship", WhseShptLine."Qty. Picked" - WhseShptLine."Qty. Shipped");
         WhseShptLine."Qty. to Ship (Base)" := WhseShptLine."Qty. Picked (Base)" - WhseShptLine."Qty. Shipped (Base)";
-        WhseShptLine.Status := WhseShptLine.CalcStatusShptLine();
+        WhseShptLine.Status := WhseShptLine.GetShipmentLineStatus();
         OnBeforeWhseShptLineModify(WhseShptLine, GlobalWhseActivLine, WhseActivityLineGrouped);
         WhseShptLine.Modify();
         OnAfterWhseShptLineModify(WhseShptLine);
@@ -773,6 +797,20 @@ codeunit 7307 "Whse.-Activity-Register"
         OnBeforeProdCompLineModify(ProdCompLine, WhseActivityLine);
         ProdCompLine.Modify();
         OnAfterProdCompLineModify(ProdCompLine);
+    end;
+
+    local procedure UpdateProdOrderLine(WhseActivityLine: Record "Warehouse Activity Line")
+    begin
+        ProdOrderLine.Get(WhseActivityLine."Source Subtype", WhseActivityLine."Source No.", WhseActivityLine."Source Line No.");
+        ProdOrderLine."Qty. Put Away (Base)" :=
+          ProdOrderLine."Qty. Put Away (Base)" + WhseActivityLine."Qty. to Handle (Base)";
+        if WhseActivityLine."Qty. per Unit of Measure" = ProdOrderLine."Qty. per Unit of Measure" then
+            ProdOrderLine."Qty. Put Away" := ProdOrderLine."Qty. Put Away" + WhseActivityLine."Qty. to Handle"
+        else
+            ProdOrderLine."Qty. Put Away" :=
+              Round(ProdOrderLine."Qty. Put Away" + WhseActivityLine."Qty. to Handle (Base)" / WhseActivityLine."Qty. per Unit of Measure");
+        ProdOrderLine."Put-away Status" := ProdOrderLine.GetLineStatus();
+        ProdOrderLine.Modify();
     end;
 
     local procedure UpdateAssemblyLine(WhseActivityLine: Record "Warehouse Activity Line")
@@ -1221,7 +1259,7 @@ codeunit 7307 "Whse.-Activity-Register"
                         end;
                 end;
 
-                OnRegisterWhseItemTrkgLineOnAfterSetDueDate(WhseActivLine2, DueDate);
+                OnRegisterWhseItemTrkgLineOnAfterSetDueDate(WhseActivLine2, DueDate, QtyToRegisterBase);
 
                 if WhseActivLine2."Activity Type" = WhseActivLine2."Activity Type"::"Invt. Movement" then
                     case WhseActivLine2."Source Type" of
@@ -1619,6 +1657,10 @@ codeunit 7307 "Whse.-Activity-Register"
             SubTotalBase :=
               QtyInWhseBase -
               QtyOnPickBinsBase - QtyOnOutboundBinsBase - QtyOnDedicatedBinsBase;
+
+            if (WhseActivLine."Action Type" = WhseActivLine."Action Type"::Take) and (WhseActivLine."Bin Code" = Location."Shipment Bin Code") then
+                SubTotalBase += QtyOnOutboundBinsBase;
+
             if WhseActivLine."Activity Type" <> WhseActivLine."Activity Type"::"Invt. Movement" then
                 SubTotalBase -= Abs(Item."Reserved Qty. on Inventory");
 
@@ -1651,8 +1693,6 @@ codeunit 7307 "Whse.-Activity-Register"
             end else
                 TotalAvailQtyBase := QtyOnPickBinsBase;
         end else begin
-            ItemLedgEntry.SetCurrentKey(
-              "Item No.", Open, "Variant Code", Positive, "Location Code", "Posting Date", "Expiration Date", "Lot No.", "Serial No.");
             ItemLedgEntry.SetRange("Item No.", WhseActivLine."Item No.");
             ItemLedgEntry.SetRange("Variant Code", WhseActivLine."Variant Code");
             ItemLedgEntry.SetRange(Open, true);
@@ -1666,7 +1706,7 @@ codeunit 7307 "Whse.-Activity-Register"
 
             LineReservedQtyBase :=
                 WhseAvailMgt.CalcLineReservedQtyOnInvt(
-                    WhseActivLine."Source Type", WhseActivLine."Source Subtype", WhseActivLine."Source No.", WhseActivLine."Source Line No.", WhseActivLine."Source Subline No.", false, TempWhseActivLine2);
+                    WhseActivLine."Source Type", WhseActivLine."Source Subtype", WhseActivLine."Source No.", WhseActivLine."Source Line No.", WhseActivLine."Source Subline No.", false, WhseItemTrackingSetup, WhseActivLine);
 
             TotalAvailQtyBase :=
               QtyInWhseBase -
@@ -2151,6 +2191,7 @@ codeunit 7307 "Whse.-Activity-Register"
         TempWarehouseActivityLine.SetRange("Action Type", WarehouseActivityLine."Action Type");
         TempWarehouseActivityLine.SetRange("Original Breakbulk", WarehouseActivityLine."Original Breakbulk");
         TempWarehouseActivityLine.SetRange("Breakbulk No.", WarehouseActivityLine."Breakbulk No.");
+        TempWarehouseActivityLine.SetRange("Unit of Measure Code", WarehouseActivityLine."Unit of Measure Code");
         if TempWarehouseActivityLine.FindFirst() then begin
             TempWarehouseActivityLine."Qty. to Handle" += WarehouseActivityLine."Qty. to Handle";
             TempWarehouseActivityLine."Qty. to Handle (Base)" += WarehouseActivityLine."Qty. to Handle (Base)";
@@ -2672,7 +2713,7 @@ codeunit 7307 "Whse.-Activity-Register"
     end;
 
     [IntegrationEvent(false, false)]
-    local procedure OnRegisterWhseItemTrkgLineOnAfterSetDueDate(WarehouseActivityLine: Record "Warehouse Activity Line"; var DueDate: Date)
+    local procedure OnRegisterWhseItemTrkgLineOnAfterSetDueDate(WarehouseActivityLine: Record "Warehouse Activity Line"; var DueDate: Date; var QtyToRegisterBase: Decimal)
     begin
     end;
